Add phosh_private_xdg_switcher protocol
Simple private protocoll to list and raise toplevels by app_id and
title. We wanted a minimal set of objects, requests and events. Since
this is only meant for the app overview page in the shell we can take
many shortcuts and leave out e.g. XWayland for the moment.

This will be replaced by wlr_foreign_toplevel_unstable_v1 later on.